+++ NEWS.md
@@ -1,3 +1,4 @@
 # varutils 0.1.0
 
-* Development version, not yet tested nor CRAN suitable
+* 11/10/2023 - Added testthat (& vdiffr) basic tests
+* 11/06/2023 - Initial update to github (development version, not yet tested nor CRAN suitable)